V10.010.9.4k64
Dear Users,

1. The way of memory allocation in the MAIN level follows that in FFS
for future extention.

2. Some infinite loop at some errors in MAIN are modified to a simple
exit of SAD, as there is no way to rescue it.

3. The calculation of Effective Voltage, Effective hHarmonic Number is
based on V(z), dV(z)/dz, d^2V(z)/dz^2. The estimation of the
equilibrium position is improved, esp. when there are many different rf
frequencies in a rin.

4. The internal scaling of parameters in OYStability is modified.

(Packages only)
5. ListQ was defined. It has been used in some plots, but ineffective.

6. Positive, Negative, NonPositive, and NonNegative becomes Listable.

7. The legend box in plots becomes weaker and the level is lowered
below plots.

8. ItemCGet of Canvas is implemented correctly.

9. The default size of Canvas can be specified Default$CanvasWidth and
Default$CanvasHeight, The default of Default$CanvasWidth is 1024 and
Default$Canvasheight Default$CanvasWidth*768/1024.