V1.0.10.9.5k64
1. The effect of F1 on the emittance calculation of BEND and QUAD has
been changed. Now it the end of magnet touches to the other BEND or
QUAD with separation of zero thickness, the F1 of that end is ignored
on the emittance calculation. The effect on focusing remains. MULT has
not been taking F1 into account for the emittance calculation.

2. A bug to crash SAD of afsad in expln has been fixed.

3. TMACRO.inc has been changed to a module tmacro.

4. Modules ffs, ffs_pointer, ffs_flag, ffs_fit have been defined to
simplify the interface to several optics/matching functions.

5. The closed orbit calculation in emittance has been modified.