V1.0.10.11k64-pre3
1. The next version V1.0.10.11k64-pre3 will be able to calculate 6D
optics for a linac (i.e. the design momentum depends on s).

2. The MATRIX(MAT) command and the TransferMatrix function will return
6x6 matrix when CALC6D is ON. TransferMatrix takes a new option
Symplectic, which returns a symplectic matrix for a linac. The default
is Symplectic->False, which means a physical transfer matrix, taking
the chenge of the design momentum due to s.

3. A bug in calculation of dispersions for CALC4D is corrected. This
bug was introduced in V1.0.10.11k64-pre to make the results
unpredictable in a case of linac.