V1.1.0.3k64
1. The next version 1.1.0.3k64 brings a few new options in Graphics:
FrameFontScale, LegendFontScale. These specify a relative sizes of
FrameLabel and Legend. FrameFontScale (also TickFontScale) accepts a
real number or a list of real numbers. If it is a real number, it is
applied to all frames. If it is a list, it is applied in the order of
bottom, left, top, right in the same way as FrameLabel. If its length
is shorter than 4, 1's are supplemented to the right.

2. OpticsPlot accepts an option RemoveOverlap. If it is not "L$NAME",
the overlapping of the labels of the lattice is untouched.