V1.1.0.10k64
Bug Fixes:
1. The value of bucket height by EMIT has been wrong since the
introduction of RADTAPER about 2 years ago.

Changes:
2. With RADTAPER, FSHIFT is automatically updated to cancel the
difference in the path length. The momentum deviation of the closed
orbit at the entrance is always kept zero.

3. RF-related parameters in EMIT, such as SIGZ, are calculated even
with NORFSW, by assuming the total voltage of CAVI and MULT. This
approximation is not bad unless the synchrotron tune is large nor there
are variation in phases and frequencies in cavities, nor dispersions at
cavities.

4. The closed orbit finding routine for EMIT is somewhat improved to
reduce the necessity of setting the RF phase manually.