V1.1.1k64pre
Experimental:
1. Now keywords for MULT can take a list to represent the longitudinal
profile of the parameters. Usage:
LINE[“L”, “Q1”]={0.2, 0.6, 0.2];
LINE[“K1”, “Q1”]={0.01, 0.05, 0.01};
Remarks:
- The keyword “L” must be given as a List to define the length of each
segment. Other keywords are either a Real or a List with the same
number of components as “L”.
- If a key is a list, the value corresponding to the segment for “L” is
taken. otherwise the Real value is used commonly for all segments.
- For a matching, the sum of the segments are used, and the resulting
value is distributed to all segments, keeping the same profile.
- Commands SAVE and RESET should work as expected.
- In commands VAR, DISP, and element-value commands (such as Q1 K1
0.02) , the summation of parameters are displayed.
- TYPE shows the Listed parameters as List, but it is not readable by
MAIN yet.
- For LINE and Element functions, keyword$SUM returns the sum of the
List if the value of keyword is a List. Otherwise it returns the same
value as keyword. Example: LINE[“K1$SUM”, ”Q1”]. Setting to keyword$SUM
is not yet implemented as expected.
- Currently the List keyword work only for MULT. However, it is not
inhibited to set to other elements, which will cause unpredictable
results.
- OpticsPlot and GeometryPlot should work properly.

2. Functions Module and With now accept more than 2 arguments. In such
a case, Module[l_, f__] means Module[l, Module[f]]. Same for With. This
makes it convenient to define local symbols cumulatively.