V1.1.1k64_pre07
Experimental:
1. The way of longitudinal profiling of MULT has been changed. Now a
new keyword PROFILE is defined and the profile of all keywords are
stored under PROFILE. The value is a list of the form:

{{key1, {v11,..,v1n}, {key2, {v21,...,v2n}}, ...}

where key1, key2, ... are the keywords of MULT. The keyword "L" must be
one of them. The values of profiles v11,.. vmn are real numbers. All
keys must have an equal length of the values.
 The actual value of each segment is given by the product of profile
values and the corresponding values of each key, ie., the "master
value". There is no automatic normalization of the values of profile.

Example:
 Element["PROFILE","Q1"]={{"L",{0.1,0.7,0.2}},{"K1",{0.15,0.6,0.25}}};
 TYPE Q1;
 ;
 MULT   Q1      =(L =1    K1 =.1   PROFILE = {{"L", {.1, .7, .2}},
{"K1", {.15, .6, .25}}} )
 ;

2. A few utilities are defined for Element for handling the profile:

Element["PROFILE"[key], e] returns the list of profile values for key.
Element["PROFILE"[key], e] = {v1,...,vn} sets the value of profile for
key, keeping the profiles for other keys.
Element["PROFILE", e]={key, {v1,..., vn}} is same as above.

Remarks:
a. The master values of each key must be a Real.
b. The geometry and path lengths are calculated based on the master
values.
c. The output by TYPE cannot be read by MAIN yet.
d. The treatment of fringes for intermediate points is not correct yet.
e. Further customizations of the way to determine the segment values
from the master values and the profile will be made depending on
demands.
===========

Changes:
1. ListContourPlot and ListDensityPlot now take a grid with unequal
spacings. it is specified by MeshRange -> {x_List, y_List} , where x
and y can be lists to specify the values at each grid. Their lengths
must be equal to the Dimensions of the data.

2. A new color function Green is defined for ListContourPlot and
ListDensityPlot, invoked by ColorFunction -> Green.

3. Fine adjustments are done for the alignment of FrameLabels.