V1.1.1k64_pre09
experimental:
1. PROFILE of MULT now accepts a combination of {key, master} as the
specification of a key. key_String and {key_String} is equivalent to
{key, key}. The value of each segment is given by the value of master
times the value of the corresponding segment of the profile. If the key
has profiles with multiple masters, all contributions are added.

2. If the key of MULT is a integrated quantity, and the PROFILE is
given, the value of the key is given by the value of the key * (the
segment of PROFILE of “L”), unless the profile with {key, key} is
defined. The contributions from the profiles of the key are added. The
integrated keys are “L”, “ANGLE”, “VOLT, “DVOLT”, “K0” to “K21”, and
“SK0” to “SK21”.

Bug fixes:
1. Pattern matching with duplicated pattern symbols could cause wrong
results in some cases.

2. The tag of the right vertical frame line of a plot has been
incorrect.

Changes:
1. The internal coding has been changed in many places.