V1.1.1k64
Bug fixes:
1. The transfer matrix of BEND and MULT with nonzero ANGLE and nonzero
K1 has been wrong since 1.0.10.9.1k64, on around Mar. 14, 2016. The
amount of error was usually not large unless the horizontal focusing
(1/rhob + K1) is too weak.

2. DynamicApertureSurvey with artificial damping has not been working
for a while, probably within a half year.

Changes:
1. New functions FitGaussian and FitEmit are implemented:
FitGaussian performs Gaussian fit of 1D list data:

FitGaussian[data, [opt, ...]]

returns a list

{sigma, mean, {sigma_conf, mean_conf}, chisq} ,

where sigma_conf and mean_conf are the confidence interval of the
results sigma and mean, respectively.
opt is options to Fit. If Plot->True is specified, a plot of the
fitting will be made.

FitEmit obtains the emittance, beta, alpha, etc. from data of particles
x and px using FitGaussian.

FitEmit[x, px] ,

where x and px are the lists of data of the particles in the phase
space, returns a list

{{xmean, pxmean, alpha, beta, emittance},
{xmean, pxmean, alpha, beta, emittance}_conf},

where the second component is the confidence interval of the first.

2. The character "\m" (RETURN) is added to WordSeparators to read a
file generated by Windows more easily. Now the default WordSeparators
are " ,\t\m".