V1.1.1.2k64
Changes:

1. The synchrotron radiation in tracking with RING, RAD and FLUC, now
take the real spectrum of photons. Historically, The real spectrum has
been included for TRPT, not for RING. It is changed. Now the real
spectrum considering the probability of photon emission are all taken
into account. However, unlike TRPT, for RING, the stochasticity of the
emission point along the orbit is not yet simulated. The result will
converge to the real value when EPS become sufficiently small.

2. Now synchrotron radiation in MULT with higher multipoles as well as
fringe field of all magnets contribute to the synchrotron radiation.
They are calculated via kinematics, not from the fields.

3. The settings of OpticsProlog and OpticsEpilog are reset for USE and
VISIT. For VISIT, they are restored by BYE.

4. A new variable NFAMP is defined to specify the phase angle of
particles for MatchingAmplitude. The defalut is 3, which gives the same
setting as before. Any positive number should work. If it is zero,
MatchingAmplitude is turned off.

Bug fixes:

1. Matching with CALC6D/TRPT/INS have not been working properly, due to
miscalculation of the phase advances in the middle of matching.